feat: add evaluation logic and eror handeling
The calculator now wont evaluate more than a single pair of numbers at a
time. Long decimal answers are rounded. Added a justEvaluated check so
pressing "=" does not cause problems when there is nothing to evaluate
and the display is cleared when pressing a new number after evaluation.
We display a snarky error message when trying to divide by zero and we
don't allow more than one decimal in a number, still need to add
feedback about that.